+++ gdtoolkit/formatter/expression_utils.py
@@ -47,13 +47,15 @@ def is_trailing_comma(expression: Node) -> bool:
 def is_expression_forcing_multiple_lines(
     expression: Node, standalone_comments: List[Optional[str]]
 ) -> bool:
-    if has_trailing_comma(expression):
-        return True
-    if _is_multiline_string(expression):
+    if has_trailing_comma(expression) or _is_multiline_string(expression):
         return True
     if isinstance(expression, Token):
         return False
-    if _has_standalone_comments(expression, standalone_comments):
+    if (
+        _has_standalone_comments(expression, standalone_comments)
+        or _is_multistatement_lambda(expression)
+        or _is_unistatement_lambda_with_compoud_statement(expression)
+    ):
         return True
     for child in expression.children:
         if is_expression_forcing_multiple_lines(child, standalone_comments):
@@ -82,10 +84,30 @@ def _is_multiline_string(expression: Node) -> bool:
 
 def _has_standalone_comments(
     expression: Tree, standalone_comments: List[Optional[str]]
-):
+) -> bool:
     return any(
         comment is not None
         for comment in standalone_comments[
             get_line(expression) : get_end_line(expression)
         ]
+    )
+
+
+def _is_multistatement_lambda(expression: Tree) -> bool:
+    return (
+        isinstance(expression, Tree)
+        and expression.data == "lambda"
+        and len(expression.children) > 2
+    )
+
+
+# TODO: remove once such statements are supported
+def _is_unistatement_lambda_with_compoud_statement(expression: Tree) -> bool:
+    return (
+        isinstance(expression, Tree)
+        and expression.data == "lambda"
+        and len(expression.children) == 2
+        and isinstance(expression.children[1], Tree)
+        and expression.children[1].data
+        in ["if_stmt", "while_stmt", "for_stmt", "for_stmt_typed", "match_stmt"]
     )