+++ tests/formatter/test_scripts_validity.py
@@ -11,8 +11,6 @@ DATA_DIR = "./input-output-pairs"
 EXCEPTIONS = set(
     [
         # cases where Godot does more than just parsing
-        "simple_inline_lambdas.in.gd",
-        "simple_inline_lambdas.out.gd",
         "inline_lambdas_w_comments.in.gd",
         "inline_lambdas_w_comments.out.gd",
         "long_inline_lambdas.in.gd",